Cinema
This website is built by .NET C# programming language along with JS, HTML, CSS and SQL. The website can help users book a seat for their preferred movie cchoice from the ones being shown in the cinema. Their is also and admin that can add, delete movies according to categories such as action, comedy and etc. and supervise seats booked so as to avoid conflict.